+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 ## [Unreleased]
 ### Added
 * Add searchbox filter on trackers table (`OSIDB-4062`)
+* Show CVE ID on page title (`OSIDB-3715`)
+
+### Changed
+* Renamed Incident states (`OSIDB-3850`)
+
+## [2025.4.0]
+### Added
+* Validate Discouraged an Prohibited CWE usage (`OSIDB-4127`)
+
+### Fixed
+* Fixed inconsistent local storage cached values (`OSIDB-4129`)
+* Added CVE ID link on flaw form (`OSIDB-3118`) 
 
 ## [2025.3.2]
 ### Added
@@ -434,7 +446,8 @@ The first release for user testing, briefly reaching feature parity with OSIDB
 ### Added
 * Early repo layout & Flaw template
 
-[Unreleased]: https://github.com/RedHatProductSecurity/osim/compare/v2025.3.2...HEAD
+[Unreleased]: https://github.com/RedHatProductSecurity/osim/compare/v2025.4.0...HEAD
+[2025.4.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.3.2...v2025.4.0
 [2025.3.2]: https://github.com/RedHatProductSecurity/osim/compare/v2025.3.1...v2025.3.2
 [2025.3.1]: https://github.com/RedHatProductSecurity/osim/compare/v2025.3.0...v2025.3.1
 [2025.3.0]: https://github.com/RedHatProductSecurity/osim/compare/v2025.2.0...v2025.3.0